Made initial restructuring code fix for QuestionnaireActivity and hooked up the QuestionnaireViewModel to the activity and relevant Dao
